refactor(core): keep provisioning helpers generic
